backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
coverage==7.8.0
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.7.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
idna==3.10
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
